strkey/internal/crc16: move crc16 inside internal
